+++ autoopts/enumeration.c
@@ -1,6 +1,6 @@
 
 /*
- *  $Id: enumeration.c,v 3.7 2003/02/16 00:04:40 bkorb Exp $
+ *  $Id: enumeration.c,v 3.8 2003/03/02 23:39:10 bkorb Exp $
  *
  *   Automated Options Paged Usage module.
  *
@@ -107,9 +107,9 @@ optionEnumerationVal( pOpts, pOD, paz_names, name_ct )
     tCC**     paz_names;
     int       name_ct;
 {
-    size_t  len;
-    int     idx;
-    int     res = -1;
+    size_t      len;
+    int         idx;
+    intptr_t    res = -1;
 
     /*
      *  IF the program option descriptor pointer is invalid,
@@ -134,7 +134,7 @@ optionEnumerationVal( pOpts, pOD, paz_names, name_ct )
         /*
          *  Replace the enumeration value with the name string.
          */
-        return (char*)paz_names[ (int)(pOD->pzLastArg) ];
+        return (char*)paz_names[ (intptr_t)(pOD->pzLastArg) ];
 
     default:
         break;
